pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jetshift"
version = "1.0.0-alpha.1"
description = "JetShift ETL Tool"
authors = [{ name = "Md Obydullah", email = "hi@obydul.me" }]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "django==5.1.7",
    "channels==4.2.0",
    "uvicorn[standard]==0.34.0",
    "daphne==4.1.2",
    "djangorestframework==3.15.2",
    "djangorestframework-simplejwt==5.5.0",
    "django-cors-headers==4.6.0",
    "mysqlclient==2.2.7",
    "psycopg2==2.9.10",
    "channels-redis==4.2.1",
    #    "jetshift-core==1.0.0a32",
    "clickhouse-sqlalchemy==0.3.2",
    # jetshift-core (previously used via PyPI)
    "prefect==3.2.14",
    "pymysql==1.1.1",
    "clickhouse-driver==0.2.9",
    "requests==2.32.3",
    "pandas==2.2.2",
    "sqlalchemy==2.0.39",
    "faker==27.4.0",
    "click==8.1.8",
    "pyyaml==6.0.2",
    "django-debug-toolbar==5.2.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["jetshift_core*", "app*", "jetshift*"]
exclude = ["tests*"]

[project.scripts]
jetshift = "jetshift_core.commands.main:main"
